+++ src/main/java/com/hoangluongtran0309/releaseflow/change/InvalidChangeFilterException.java
@@ -1,0 +1,8 @@
+package com.hoangluongtran0309.releaseflow.change;
+
+public class InvalidChangeFilterException extends RuntimeException {
+
+    public InvalidChangeFilterException(String message) {
+        super(message);
+    }
+}